backend/requirements.txt
Django==5.0.6
asgiref
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
# for swagger docs
drf-spectacular
pyyaml
uritemplate
inflection